refactor: isolate bungee & velocity classes
