+++ src/main/java/dev/lone/bungeepackfix/bungee/libs/YamlConfig.java
@@ -11,7 +11,7 @@
  * License: Apache 2.0
  * Licensor: LoneDev
  */
-package dev.lone.bungeepackfix.libs;
+package dev.lone.bungeepackfix.bungee.libs;
 
 import net.md_5.bungee.api.plugin.Plugin;
 import net.md_5.bungee.config.Configuration;
